test/agents/requirements.txt
livekit-agents>=1.0.0
livekit-plugins-deepgram>=1.0.0
livekit-plugins-elevenlabs>=1.0.0
livekit-plugins-openai>=1.0.0
livekit-plugins-cartesia>=1.0.0
livekit-plugins-silero>=1.0.0
livekit-plugins-turn-detector>=1.0.0
python-dotenv~=1.0
